fix(engine): strengthen deterministic hashing across doctor preflight, proof ingestion, runtime gate, and timescale certificates
Comprehensive hashing correctness pass — every content hash must
include all semantically meaningful fields to prevent collisions
between structurally different objects:

- react_doctor_preflight.rs: Introduce length-prefixed hash helpers
  (hash_u64, hash_str, hash_string_vec, hash_string_set) to prevent
  concatenation collisions. DoctorCheck::content_hash() now includes
  remediation, related_mismatch_ids, and tags. GuidanceEntry::
  content_hash() now includes priority, category, severity, and
  originating_check_ids. run_preflight result hash includes
  entries_analyzed count. build_support_bundle hash now includes entry
  categories, all summary counters, aggregate_score, is_ready, and
  guidance hashes. GuidanceEntry.guidance_hash is now computed from
  the full payload via content_hash() instead of just id+title.
  Add tests: doctor_check_content_hash_changes_when_payload_changes,
  preflight_result_hash_changes_when_entries_analyzed_changes,
  guidance_content_hash_changes_when_metadata_changes,
  generate_guidance_sets_hash_from_full_payload,
  support_bundle_hash_changes_when_guidance_payload_changes.

- proof_ingestion.rs: ProofInput::canonical_bytes() now includes
  issuer_signature and payload. Enum fields use to_string() instead
  of `as u8` cast (which silently truncates discriminant values).
  OptimizerHypothesis::canonical_bytes() includes optimization_class
  debug repr and uses string-based enum representation for kind/risk.

- runtime_comparison_gate.rs: Evidence hash now includes blocker
  count and reproducibility entry count so two runs with different
  blockers but same benchmark totals no longer collide.

- timescale_separation_certificate.rs: Fix potential u64 overflow
  in timescale ratio computation by widening operands to u128 before
  the multiplication by 1_000_000 — without this, two large u64
  observation counts could silently wrap, producing incorrect ratios.

+++ crates/franken-engine/src/timescale_separation_certificate.rs
@@ -423,17 +423,17 @@ pub fn compute_timescale_ratio(
     let obs_ratio = if obs_a == 0 || obs_b == 0 {
         0
     } else if obs_a >= obs_b {
-        obs_a * 1_000_000 / obs_b
+        (obs_a as u128 * 1_000_000 / obs_b as u128) as u64
     } else {
-        obs_b * 1_000_000 / obs_a
+        (obs_b as u128 * 1_000_000 / obs_a as u128) as u64
     };
 
     let write_ratio = if write_a == 0 || write_b == 0 {
         0
     } else if write_a >= write_b {
-        write_a * 1_000_000 / write_b
+        (write_a as u128 * 1_000_000 / write_b as u128) as u64
     } else {
-        write_b * 1_000_000 / write_a
+        (write_b as u128 * 1_000_000 / write_a as u128) as u64
     };
 
     let (ratio, basis) = if obs_ratio <= write_ratio {
